nails-backend-main/requirements.txt
amqp==5.0.9
aniso8601==7.0.0
anyascii==0.3.2
anytree==2.8.0
asgiref==3.4.1
beautifulsoup4==4.11.2
billiard==3.6.4.0
boto3==1.20.26
botocore==1.23.26
celery==5.2.3
certifi==2021.10.8
charset-normalizer==3.1.0
click==8.0.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.4
cssselect==1.2.0
defusedxml==0.7.1
Deprecated==1.2.13
diff-match-patch==20200713
Django==3.2.10
django-celery-beat==2.2.1
django-celery-results==2.2.0
django-clearcache==1.2.1
django-cors-headers==3.10.1
django-debug-toolbar==3.2.4
django-editorjs-fields==0.2.7
django-filter==21.1
django-graphql-jwt==0.3.4
django-import-export==3.2.0
django-js-asset==2.0.0
django-modelcluster==6.0
django-mptt==0.14.0
django-permissionedforms==0.1
django-redis==5.2.0
django-storages==1.12.3
django-taggit==3.1.0
django-timezone-field==4.2.1
django-treebeard==4.6.1
django-versatileimagefield==3.0
djangorestframework==3.14.0
draftjs-exporter==2.1.7
et-xmlfile==1.1.0
flower==1.0.0
graphene==2.1.9
graphene-django==2.15.0
graphene-django-filter==0.6.4
graphene-django-pagination @ https://github.com/instruct-br/graphene-django-pagination/archive/master.zip
graphql-core==2.3.2
graphql-relay==2.0.1
gunicorn==20.1.0
html5lib==1.1
humanize==3.13.1
idna==3.4
jmespath==0.10.0
kombu==5.2.3
l18n==2021.3
lxml==4.9.3
markdown2==2.4.10
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.1.2
packaging==21.3
Pillow==8.4.0
prometheus-client==0.12.0
promise==2.3
prompt-toolkit==3.0.24
psycopg2-binary==2.9.3
PyJWT==2.6.0
pyparsing==3.0.6
pyquery==2.0.0
python-crontab==2.6.0
python-dateutil==2.8.2
python-magic==0.4.27
pytz==2021.3
PyYAML==6.0
readtime==3.0.0
redis==4.1.0
requests==2.28.2
Rx==1.6.3
s3transfer==0.5.0
sentry-sdk==1.5.1
singledispatch==4.0.0
six==1.16.0
soupsieve==2.4
sqlparse==0.4.2
stringcase==1.2.0
tablib==3.4.0
telepath==0.3
text-unidecode==1.3
tornado==6.1
tzdata==2021.5
urllib3==1.26.7
vine==5.0.0
wcwidth==0.2.5
webencodings==0.5.1
Willow==1.4.1
wrapt==1.15.0
xlrd==2.0.1
xlwt==1.3.0
